moves constant imported in layout component config to own file to avoid css loading error
