+++ src/app-components/panel/constants.ts
@@ -1,0 +1,6 @@
+export const PANEL_VARIANTS = {
+  Info: 'info',
+  Warning: 'warning',
+  Error: 'error',
+  Success: 'success',
+} as const;
